refactor: split napi transform bindings
